+++ sdk/monitor/azure-monitor-query/samples/sample_logs_query_key_value_form.py
@@ -21,14 +21,13 @@ end_time = datetime.now(UTC())
 # returns LogsQueryResult 
 response = client.query(os.environ['LOG_WORKSPACE_ID'], query, duration=timedelta(days=1), end_time=end_time)
 
-if not response.tables:
-    print("No results for the query")
-
-for table in response.tables:
-    pd.json_normalize
+try:
+    table = response.tables[0]
     df = pd.DataFrame(table.rows, columns=[col.name for col in table.columns])
     key_value = df.to_dict(orient='records')
     print(key_value)
+except TypeError:
+    print(response.error)
 
 """
 [